+++ packages/opentelemetry-resource-detector-aws/src/detectors/AwsEc2Detector.ts
@@ -53,16 +53,14 @@ class AwsEc2Detector implements Detector {
       region,
       availabilityZone,
     } = await this._awsMetadataAccessor();
-    return new Resource(
-      {
-        [CLOUD_RESOURCE.PROVIDER]: 'aws',
-        [CLOUD_RESOURCE.ACCOUNT_ID]: accountId,
-        [CLOUD_RESOURCE.REGION]: region,
-        [CLOUD_RESOURCE.ZONE]: availabilityZone,
-        [HOST_RESOURCE.ID]: instanceId,
-        [HOST_RESOURCE.TYPE]: instanceType,
-      },
-    );
+    return new Resource({
+      [CLOUD_RESOURCE.PROVIDER]: 'aws',
+      [CLOUD_RESOURCE.ACCOUNT_ID]: accountId,
+      [CLOUD_RESOURCE.REGION]: region,
+      [CLOUD_RESOURCE.ZONE]: availabilityZone,
+      [HOST_RESOURCE.ID]: instanceId,
+      [HOST_RESOURCE.TYPE]: instanceType,
+    });
   }
 
   /**
